requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
e2b==1.5.1
e2b-code-interpreter==1.5.1
email-validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
frozenlist==1.7.0
fsspec==2024.12.0
google-search-results==2.4.2
grpcio==1.73.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.1
hyperframe==6.1.0
idna==3.10
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
json-repair==0.39.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-mongodb==0.6.2
langchain-neo4j==0.4.0
langchain-openai==0.3.23
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.9
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-checkpoint-mongodb==0.1.4
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.4.8
lark==1.2.2
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
motor==3.7.1
mpmath==1.3.0
multidict==6.6.3
mypy-extensions==1.1.0
neo4j==5.28.1
neo4j-graphrag==1.7.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.2
openai==1.86.0
orjson==3.11.1
ormsgpack==1.10.0
packaging==24.2
pandas==2.3.1
pillow==11.2.1
pinecone==7.3.0
pinecone-plugin-assistant==1.7.0
pinecone-plugin-interface==0.0.7
portalocker==2.10.1
propcache==0.3.2
protobuf==5.29.5
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.1
pymongo==4.12.1
pypdf==5.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
qdrant-client==1.14.3
redis==6.2.0
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.7
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
sentence-transformers==5.0.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.41
starlette==0.46.2
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.54.0
typer==0.16.0
types-pyyaml==6.0.12.20250516
typing-extensions==4.14.1
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
